Department.js improved , folder structure improved
