+++ client/src/components/teams2/portfolio/Content.js
@@ -1,7 +1,7 @@
 import React, { useState } from 'react';
 import { motion, AnimatePresence } from 'framer-motion';
 import Template from '../Template';
-import Department from './Department';
+import Department from '../Department';
 import { Users, UserCheck } from 'lucide-react';
 
 const Content = () => {
@@ -27,7 +27,7 @@ const Content = () => {
       variants={containerVariants}
       className="min-h-screen bg-zinc-50"
     >
-      <Department tag="CONTENT TEAM" />
+      <Department tag="CONTENT TEAM" des="The storytellers behind every impactful narrative"/>
       
       <div className="container mx-auto px-4 py-12">
         {/* Toggle Section */}
