+++ client/src/components/teams2/portfolio/Marketing.js
@@ -1,7 +1,7 @@
 import React, { useState } from 'react';
 import { motion, AnimatePresence } from 'framer-motion';
 import Template from '../Template';
-import Department from './Department';
+import Department from '../Department';
 import { Users, UserCheck } from 'lucide-react';
 
 const Marketing = () => {
@@ -27,7 +27,7 @@ const Marketing = () => {
       variants={containerVariants}
       className="min-h-screen bg-zinc-50"
     >
-      <Department tag="MARKETING AND SPONSORSHIP TEAM" />
+      <Department tag="MARKETING AND SPONSORSHIP TEAM" des="The strategists behind our growth and partnerships"/>
       
       <div className="container mx-auto px-4 py-12">
         {/* Toggle Section */}
